+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+**This version changed some mappings in presets, so the same input probably won't return the same result as version 1.0.0.**
+
+- Add more word mappings, so it's on par with mohan-cao's latest owoify.
+- Add uwuify and uvuify for convenience.
+
 ## 1.0.0
 - **(Breaking change)** Change owoness level from strings to symbols.
 - Add proper RBS type signature to internal function(s).